# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDesk Remote API
  version: 0.1.0
  description: REST API skeleton for OpenDesk Remote control plane.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed session bearer token backed by a server-side session hash, or user-bound OpenDesk API token.
    sessionCookie:
      type: apiKey
      in: cookie
      name: opendesk_session
paths:
...
  /api/v1/devices:
    get:
      summary: List devices
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Devices returned }
        "401": { description: Authentication required }
    post:
      summary: Create device
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "201": { description: Device created }
        "401": { description: Authentication required }
...
  /api/v1/device-groups:
    get:
      summary: List device groups
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Device groups returned }
        "401": { description: Authentication required }
    post:
      summary: Create device group
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                member_device_ids:
                  type: array
                  items: { type: integer }
      responses:
        "201": { description: Device group created }
        "400": { description: Invalid device group }
        "401": { description: Authentication required }
...
  /api/v1/devices/{id}:
    get:
      summary: Get device
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Device returned }
        "401": { description: Authentication required }
        "404": { description: Device not found }
    put:
      summary: Update device
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rustdesk_id, name]
              properties:
                rustdesk_id:
                  type: string
                name:
                  type: string
                alias:
                  type: string
                owner_user_id:
                  type: integer
                  nullable: true
                status:
                  type: string
                  enum: [online, offline, disabled]
                platform:
                  type: string
                client_version:
                  type: string
                opendesk_client_version:
                  type: string
                last_ip:
                  type: string
      responses:
        "200": { description: Device updated }
        "400": { description: Invalid device }
        "401": { description: Authentication required }
        "404": { description: Device not found }
...